Redesigning lf.cache.Journal commit behavior.
Specifically:
- Modifying behavior of insert(), insertOrReplace(), update(), remove() to
immediately update the cache and indices. Before this was happening only
when Journal#commit() was called.
- Simplifying checkPrimaryKeyExistence() check to look only into the indices.
No need for previous logic that was taking into account changes that were
not reflected in the indices. Removing obsolete logic (findExistingRowId_ and
findExistingRowIdInJournal_).
- Adding new class lf.cache.TableDiff that records all changes that have
happened on a table. Obsoleting Journal's internal data structures
changedRows and newRowIds in favor of the new more powerful TableDiff class.

TableDiff#getReverse() reverses the diff and will be used for rolling back the
changes from cache and indices in the case the journal failed to be merged to
backstore (in follow up CL).
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=77279225