Fixing lf.tree.map to correctly replicate the original tree structure.
In certain cases the resulting tree did not have the exact same
structure as the original one. The previous unit test was not catching
this error because it was checking that the in-order traversal was
visiting nodes in the same order, instead of actually comparing the
entire tree structures.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=78362691