Re-using code between QueryTask and ObserverTask.
Specifically:
- Extracting a superclass out of QueryTask.
- Moving remaining QueryTask logic to UserQueryTask subclass.
- Making ObserverTask a subclass of QueryTask and renaming to
ObserverQueryTask.
- Simplifying constructor of QueryTask to only expect an array of queries.
Before it was also expecting an array of PhysicalQueryPlan instances which
was redundant.

With this design the core logic of executing a query resides in a single place
in QueryTask, which exposes a protected onSuccess() method that subclasses can
override as needed.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=80242672