Modifying lf.index.RowId#getName to return the normalized name.
This is necessary since an IDB table will be created for a RowID
index and if the name is not normalized it will collide with other
tables RowID indices.

Also ensuring that the value returned by RowId#getName() is also
used in MemoryIndexStore to identify the RowId index. Before this
CL, getName() returned "##rowid##' but in MemoryIndexStore
"<tableName>.#" was used.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=81435565