Using indices to sort where possible.
Specifically:
- Adding a new optimization pass, OrderByIndexPass.
- The new pass transforms the tree such that indices are leveraged for sorting
where possible. It basically looks for two patterns in the tree that qualify
for this optimization, the presence of a TableAccessFullStep or an
IndexRangeScanStep node. In both cases the tree should have no
CrossProductStep or JoinStep nodes, in those cases the optimization can't be
applied.

With this optimization the OrderByIndexed performance benchmark test case executes
in ~165ms (vs ~650ms without the optimization).
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=83534035