Removing usage of Array.observe/unobserve.
It turns out that the behavior is slightly different when Array.observe
is used compared to when it is not. The former resulted in confusing
behavior, whereas the latter is the desired behavior. Combined with the
fact that there is no good reason to use Array.observe, reducing two
code paths to one, removes complexity and fixes a bug at the same time.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=84375664