Adding lf.index.Comparator#bindKeyRange method.
It is used to bind the minimum and maximum values in key ranges that are
unbound on either side, both for single and cross-column key ranges. This
is in preparation of being able to correctly handle Index#getRange() queries
for the case of cross-column indices.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=86616655