Fixing transaction commit to respect BaseTx#commit result.
Breaking BaseTx#commit to two methods, commit and commitInternal (abstract).
The latter is implemented from every subclass, and it is called by the
former one. This makes it impossible for subclasses to ignore the result
of BaseTx#commit, which was accidental (a bug).

Tests will be added in follow up CL (part of end-to-end foreign key tests).
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=97281433