Modify lf.query.SelectBuilder to match the join predicate columns with join order.
It swaps the columns in Join Predicate when they are in reverse to the order in which left outer join is performed. It also changes the join operation from LT to GT, LTE to GTE and vice versa.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=97892540