Optimizing SELECT COUNT(*).
Adding a new optimization pass that looks for SELECT COUNT(*)
queries, and modifies the execution plan to read the row count
from the RowId index instead of doing a full-table-scan and counting
the rows.

This makes SELECT COUNT(*) execution time independent of the size of a
table (it was linear to the size before).
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=98872267