Fix gulp test not working on Windows/Mac.
The common issue for both Windows and Mac is that we did not handle
fs#mkdirSync correctly. Use mkdirp package in SPAC so that we will not
encounter that error again.

For Windows, TestEnv was not creating symbolic links in the right way.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=100539589