Stop using lf.testing.getSchemaBuilder from context_test.js.
The reason is that mock schema is used by 20+ test files, and therefore it is
very hard to change it for the purposes of a single test, without affecting
other tests. A better approach (followed by this CL) is to have tests create a
schema on-the-fly, the benefits are
 1) Other tests are not affected when the schema of a single test is modified.
 2) On-the-fly schemas are minimal therefore making it easier to understand
    what is being tested.

This is in preparation of adding more tests for the case of CASCADE
constraints.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=103050011