Fix ConstraintChecker to always call keyOfIndex() before querying an index.
Before this change in some cases the value in a row's payload was directly used
to perform index queries, which even though it does not cause a bug, it is not
correct. DATE_TIME values ought to be converted to a number before being passed
to the index.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=104790239