Leveraging indices for the case of IN predicates.
If the IN predicate corresponds to an indexed column, and if the number
of values in the predicate is less than 2% of the total number of rows
in the table, then the index is leveraged to calculate the rows that
satisfy the predicate. The 2% limit, was determined by trial and error.

This optimization drops the execution time of the
SelectMultiRowIndexedInPredicate perf case from ~50ms to ~5ms.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=104908848