Changing ValuePredicate#toKeyRange() return value to lf.index.SingleKeyRangeSet.
This is in preparation of adding CombinedPredicate#toKeyRange() method for the
purposes of leveraging indices to evaluate OR predicates. This will allow
toKeyRange() to return meaningful keyRanges for the OR predicate, and keep the
signature of ValuePredicate#toKeyRange and CombinedPredicate#toKeyRange
consistent.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=105074007