Address TODO in InMemoryUpdater#updateTableIndexForRow.
The TODO is actually infeasible. The changes shall not be pushed down to
the comparator level because it will increase the time of searching a row.
Existing solution is a good compromise because it affects execution time
once per query, instead of once per row.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=105433342