Intercepting WebSql initialization error.
WebSQL does not work in Safari's incognito mode, but the error message is
confusing. Instead intercept the error convert it to an lf.Exception which
points to the FAQ.

This also fixes a case where WebSql#checkVersion_ was never fulfilling the
Promise it returned.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=105439563