Shortening perf test case names for SelectBenchmark.
The "Select" prefix is redundant since the SelectBenchmark performs only SELECT
queries. Also the "MultiRow" prefix is redundant for OR and IN queries. This
will free up some real-estate on the perf graphs by making the graph legend
smaller.

This is in preparation of adding a new test case for an OR predicate that
involves multiple columns.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=105442211